refactor(github): extract GitHub error classification to new module
- Moved classify_github_api_error, classify_by_status, and is_rate_limited functions from src/github/mod.rs to a new src/github/classify.rs module.
- Updated imports and module usage accordingly.
- Added a test helper function test_classify_error_message for integration tests.
- Improved code organization and separation of concerns for GitHub API error handling and classification.
